+++ src/mesh.cpp
@@ -9,7 +9,7 @@
 
 #define SMOOTH (v.size() > 100)
 
-bool Mesh::intersect(const Ray &r, Hit &h, float t_min)
+bool Mesh::intersect(const Ray &r, Hit &h, double t_min)
 {
     /*bool result = false;
     for (unsigned int i = 0; i < t.size(); i++)
@@ -36,7 +36,7 @@ bool Mesh::intersect(const Ray &r, Hit &h, float t_min)
     octree.intersect(r, func);
     return func.result;
 }
-bool Mesh::intersectTrig(int idx, const Ray &r, Hit &h, float t_min)
+bool Mesh::intersectTrig(int idx, const Ray &r, Hit &h, double t_min)
 {
     bool result = false;
     Triangle triangle(v[t[idx][0]], v[t[idx][1]], v[t[idx][2]], material);
